Enhancement: Added "Player Stats" window that can be toggled.
Clicking on the current Player in the top right corner of the game window will open a new window with info about player's money and properties. It updates automatically on each turn. Also fixed a bug being caused by the use of a non-thread-safe data structure (using an ArrayList to store players, now it is using a Vector).