Enhancement: Added color to player's pieces and properties.
Each player gets a fixed color at the start of the game, and this is represented in their token, owned properties, and text on the screen and in the "Player Stats" window. Also made it so the PropertyOpMenu is displayed whenever a player lands on one of their properties. Furthermore, fixed a bug with one of the lua scripts.